refactor: rename `mappings` module to `keymaps`
